requirements.txt
absl-py==0.12.0
appdirs==1.4.4
attrs==20.3.0
cached-property==1.5.2
cachetools==4.2.1
cattrs==1.0.0
certifi==2020.12.5
chardet==4.0.0
cloudpickle==1.6.0
distlib==0.3.1
filelock==3.0.12
future==0.18.2
google-auth==1.29.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
gym==0.18.0
h5py==3.2.1
idna==2.10
importlib-metadata==1.6.0
Markdown==3.3.4
more-itertools==8.3.0
numpy==1.20.2
oauthlib==3.1.0
Pillow==7.2.0
pip==20.1.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
python-include==1.0.0
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.2
setuptools==47.3.1.post20200810
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
torch==1.8.1
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.0.28
Werkzeug==1.0.1
wheel==0.34.2
zipp==3.1.0